create the product tsv file
